feat: hook useAquariumData and useSelectedAquarium
