+++ src/hooks/index.ts
@@ -1,4 +1,6 @@
-export { useAquariumDetails } from './useAquariumDetails';
+export { useAquariumData } from './useAquariumData'
 export { useCreateAquarium } from './useCreateAquarium';
+export { useLoadAquariumDetails } from './useLoadAquariumDetails';
 export { useLoadAquariums } from './useLoadAquariums';
 export { useResetParams } from './useResetParams';
+export { useSelectedAquarium } from './useSelectedAquarium';
