--- src/routes/AquariumTab/map.ts
@@ -1,23 +1,0 @@
-import { Icons } from '../../theme';
-
-
-export const Map:any = {
-  icon: {
-    'Curvo': Icons.circularShape,
-    'Sextavado': Icons.hexagonalShape,
-    'Retangular': Icons.rectangularShape,
-  },
-  pet: {
-    'Peixe': Icons.fish,
-    'Tartaruga': Icons.turtle,
-    'Cobra': Icons.snake,
-    'Sapo': Icons.frog,
-  },
-  sensor: {
-    'Temperatura': Icons.temperatureData,
-    'Saturação': Icons.saturationData,
-    'pH': Icons.phData,
-    'Nível de água': Icons.waterLevelData,
-    'Luminosidade': Icons.luminosityData,
-  },
-}
